infoとpermissionを統合:「早すぎる最適化」になっていたため
